feat: add separate component Modal
