+++ src/components/Agreement/index.ts
