+++ src/components/AvatarList/AvatarList.styled.tsx
