+++ src/components/AvatarList/AvatarList.tsx
@@ -1,35 +1,25 @@
-import { createPortal } from 'react-dom';
-import { MouseEvent, forwardRef } from 'react';
+import React, {MouseEvent} from 'react';
 import avatars from '../../avatars/avatars';
 import List from './AvatarList.styled';
-import { Backdrop, Modal } from '../AgreementModal/AgreementModal.styled';
-
-const modalRoot = document.querySelector('#modal-root');
+import Modal from '../Modal';
 
 interface IProps  {
-  onClick: (id: number) => number;
+  onClick: (id: string) => string;
   setIsAvatarList: (a: boolean) => void;
 }
 
-const AvatarList = forwardRef<HTMLInputElement, IProps>(({ onClick, setIsAvatarList }, ref) => {
-  const handleClose = (e: MouseEvent<HTMLDivElement>) => {
-    if (e.target === e.currentTarget) {
-      setIsAvatarList(false);
-    }
-  };
-  return createPortal(
-    <Backdrop ref={ref} onClick={(e: MouseEvent<HTMLDivElement>) => handleClose(e)}>
-      <Modal>
-        <List onClick={(e: MouseEvent<HTMLUListElement>) => onClick(e.target.id)}>
+const AvatarList: React.FC<IProps> = (({ onClick, setIsAvatarList }) => {
+
+  return (
+      <Modal setIsAvatarList={setIsAvatarList}>
+        <List onClick={(e: MouseEvent<HTMLUListElement>) => onClick((e.target as HTMLLIElement).id)}>
           {avatars.map((avatar, index) => (
             <li key={index}>
-              <img src={avatar} alt="Logo" id={index} />
+              <img src={avatar} alt="Logo" id={index.toString()} />
             </li>
           ))}
         </List>
       </Modal>
-    </Backdrop>,
-    modalRoot
   );
 });
